Bashify
added bash as the default shell for the splunk user
sourced /opt/splunk/bin/setSplunkEnv into /etc/bash.bashrc

bashrc runs when you:

docker exec -it splunk bash

So in a general case, nothing changes. But when you jump in to
fix/tweak, your environment is set up for you.